refactor: move the code to the root directory
