+++ .github/workflows/main.yml
@@ -26,6 +26,6 @@ jobs:
       uses: ibiqlik/action-yamllint@v1.0.0
       with:
         # File(s) or Directory, separate by space if multiple files or folder are specified
-        file_or_dir: crawler/whitelist/*.yml crawler/blacklist/*.yml
+        file_or_dir: publishers*.yml blacklist/*.yml
         # Return non-zero exit code on warnings as well as errors
         strict: true
