+++ cmd/delete.go
@@ -1,7 +1,7 @@
 package cmd
 
 import (
-	"github.com/italia/developers-italia-backend/crawler/crawler"
+	"github.com/italia/developers-italia-backend/crawler"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
