+++ crawler/crawler.go
@@ -18,10 +18,10 @@ import (
 	"time"
 
 	"github.com/alranel/go-vcsurl/v2"
-	"github.com/italia/developers-italia-backend/crawler/elastic"
-	"github.com/italia/developers-italia-backend/crawler/ipa"
-	"github.com/italia/developers-italia-backend/crawler/jekyll"
-	"github.com/italia/developers-italia-backend/crawler/metrics"
+	"github.com/italia/developers-italia-backend/elastic"
+	"github.com/italia/developers-italia-backend/ipa"
+	"github.com/italia/developers-italia-backend/jekyll"
+	"github.com/italia/developers-italia-backend/metrics"
 	httpclient "github.com/italia/httpclient-lib-go"
 	publiccode "github.com/italia/publiccode-parser-go/v3"
 	es "github.com/olivere/elastic/v7"
