+++ ipa/amministrazioni.go
@@ -15,7 +15,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/italia/developers-italia-backend/crawler/elastic"
+	"github.com/italia/developers-italia-backend/elastic"
 	es "github.com/olivere/elastic/v7"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
