+++ jekyll/recent_software.go
@@ -6,7 +6,7 @@ import (
 	"os"
 
 	"github.com/ghodss/yaml"
-	"github.com/italia/developers-italia-backend/crawler/elastic"
+	"github.com/italia/developers-italia-backend/elastic"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
 	es "github.com/olivere/elastic/v7"
