+++ whitelist/publishers.yml.example
